requirements.txt
requests==2.26.0
environs==9.3.5
python-slugify==6.1.2
redis==4.3.4
python-telegram-bot==11.1.0
geopy==2.3.0

